Resolves #23 by moving "Layer" from self._layers to self._simple_datasets
Layer files have metadata that's operated on directly, and no path to
another dataset (thus dataElement not existing for the describe object).
So, they should be considered a simple dataset, which makes the rest of
the test suite function. Thomas, if you get a chance to sign off on
this, I'd appreciate it - I think it should fix support for layer files