WIP: I'm not sure I trust these edits today until I can do better comments and ensure the test suite passes
I think that lxml may have changed a bit, sometimes raising errors for missing elements with a find rather than
returning None. So I extracted some methods to cover both cases. But I'm not positive I understand all the code yet,
and wonder if the change to use .values() on line 197 is a symptom of returning the wrong data somewhere earlier.